+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+- "/sublibrary/sublibraries": the deprecated "libraryManager" query parameter was
+  removed. It is ignored from now on; the sublibraries of the currently
+  logged-in person are always returned.
+
 ## v0.6.1
 
 - "/sublibrary/users" gained a new "idNumber" property, containing an ID number of the user, if available.